backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==1.0.4
requests==2.31.0
PyJWT==2.10.1
sib-api-v3-sdk==7.6.0
mailjet-rest==1.3.4
Flask-Compress==1.13